# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PagosDigitales API
  version: 0.0.1
  description: API unificada para Pagos Digitales con rutas dinamicas y SafetyPay Confirmation.
servers:
  - url: http://localhost:8080
paths:
  /api/v1/banks:
    post:
      summary: Consulta bancos por proveedor o todos los proveedores configurados.
      description: Si `payment_provider_code` no se envia, el backend consulta todos los proveedores activos y consolida la respuesta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BanksRequest'
      responses:
        '200':
          description: Respuesta de bancos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BanksResponse'
        '400':
          description: Request invalido o proveedor no configurado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '504':
          description: El proveedor externo excedio el timeout configurado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Error interno.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/payments:
    post:
      summary: Consulta pagos por operacion.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentsRequest'
      responses:
        '200':
          description: Respuesta de pagos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentsResponse'
        '400':
          description: Request invalido o proveedor no configurado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '404':
          description: No existe el pago solicitado en el registro interno.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '504':
          description: El proveedor externo excedio el timeout configurado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Error interno.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/direct-online-payment-requests:
    post:
      summary: Solicita un pago en linea directo.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DirectOnlinePaymentRequest'
      responses:
        '200':
          description: Respuesta de pago en linea.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectOnlinePaymentResponse'
        '400':
          description: Request invalido, proveedor no configurado o validacion de negocio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '504':
          description: El proveedor externo excedio el timeout configurado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Error interno.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/payments/notifications/merchant-events:
    post:
      summary: Recibe notificaciones de eventos del comercio y responde localmente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MerchantEventsRequest'
      responses:
        '200':
          description: Respuesta de merchant events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MerchantEventsResponse'
  /api/v1/jep/notifyPayment:
    post:
      summary: Recibe la notificacion de pago exitoso de JEPFaster.
      description: Webhook consumido por Cooperativa JEP cuando el pago QR finaliza exitosamente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JepNotifyPaymentRequest'
      responses:
        '200':
          description: Pago confirmado y actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JepNotifyPaymentResponse'
        '400':
          description: Trama invalida, estado diferente de PAGADO o error diferente de cero.
        '404':
          description: No existe el pago identificado por idtransaccion.
        '500':
          description: Error interno al procesar la notificacion.
  /api/v1/safetypay/confirmation:
    post:
      summary: Webhook de confirmaciones SafetyPay.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/SafetypayConfirmationRequest'
      responses:
        '200':
          description: CSV firmado.
          content:
            text/plain:
              schema:
                type: string
  /api/v1/issuer-commissions:
    get:
      summary: Consulta comisiones emisoras.
      description: Consulta registros de `TRX3.FEMSA_EMISOR_COMISION` y permite filtrar por `codigo_establecimiento`.
      parameters:
        - name: codigo_establecimiento
          in: query
          required: false
          schema:
            type: string
          description: Codigo de establecimiento a consultar.
      responses:
        '200':
          description: Comisiones emisoras encontradas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuerCommissionResponse'
        '400':
          description: Filtro invalido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '500':
          description: Error interno.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
  /api/v1/pagos/test:
    get:
      summary: Health check de la API.
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
components:
  schemas:
    BanksRequest:
      type: object
      properties:
        chain:
          type: integer
        store:
          type: integer
        store_name:
          type: string
        pos:
          type: integer
        payment_provider_code:
          type: integer
        channel_POS:
          type: string
        country_code:
          type: string
      required:
        - chain
        - store
        - store_name
        - pos
        - country_code
    PaymentsRequest:
      type: object
      properties:
        chain:
          type: integer
        store:
          type: integer
        store_name:
          type: string
        pos:
          type: integer
        payment_provider_code:
          type: integer
        operation_id:
          type: string
        request_datetime:
          type: string
      required:
        - payment_provider_code
        - operation_id
    DirectOnlinePaymentRequest:
      type: object
      properties:
        chain:
          type: integer
        store:
          type: integer
        store_name:
          type: string
        pos:
          type: integer
        payment_provider_code:
          type: integer
        sales_amount:
          type: object
          properties:
            value:
              type: number
            currency_code:
              type: string
        country_code:
          type: string
        bank_id:
          type: string
        merchant_set_pay_amount:
          type: boolean
        expiration_time_minutes:
          type: integer
        language_code:
          type: string
        custom_merchant_name:
          type: string
        merchant_sales_id:
          type: string
        requested_payment_type:
          type: string
        transaction_email:
          type: string
        send_email_shopper:
          type: boolean
      required:
        - payment_provider_code
        - sales_amount
        - country_code
    MerchantEventsRequest:
      type: object
      properties:
        chain:
          type: integer
        store:
          type: integer
        store_name:
          type: string
        pos:
          type: integer
        payment_provider_code:
          type: integer
        merchant_events:
          type: array
          items:
            type: object
        request_datetime:
          type: string
      required:
        - payment_provider_code
        - merchant_events
    SafetypayConfirmationRequest:
      type: object
      properties:
        ApiKey:
          type: string
        RequestDateTime:
          type: string
        MerchantSalesID:
          type: string
        ReferenceNo:
          type: string
        CreationDateTime:
          type: string
        Amount:
          type: string
        CurrencyID:
          type: string
        PaymentReferenceNo:
          type: string
        Status:
          type: string
        Signature:
          type: string
    JepNotifyPaymentRequest:
      type: object
      properties:
        idtransaccion:
          type: string
          description: Identificador generado al crear el QR.
        estado:
          type: string
          enum: [PAGADO]
        mensaje:
          type: string
          description: Mensaje del estado de la transaccion.
        identificadorsesion:
          type: string
          nullable: true
          description: Identificador de sesion opcional enviado al generar el QR.
        nummensaje:
          type: string
          description: Numero de comprobante o documento del pago.
        error:
          type: string
          enum: ['0']
      required:
        - idtransaccion
        - estado
        - mensaje
        - nummensaje
        - error
    JepNotifyPaymentResponse:
      type: object
      properties:
        status:
          type: string
          example: OK
      required:
        - status
    BanksResponse:
      type: object
      description: Respuesta normalizada de bancos.
    PaymentsResponse:
      type: object
      description: Respuesta normalizada de pagos.
    DirectOnlinePaymentResponse:
      type: object
      description: Respuesta normalizada de pagos en linea.
    IssuerCommissionResponse:
      type: object
      properties:
        codigo_establecimiento:
          type: string
          nullable: true
        total:
          type: integer
        commissions:
          type: array
          items:
            $ref: '#/components/schemas/IssuerCommissionItem'
    IssuerCommissionItem:
      type: object
      properties:
        codigo_establecimiento:
          type: string
        nombre:
          type: string
        monto_minimo:
          type: number
        comision_fija:
          type: number
        comision_variable:
          type: number
    ApiErrorResponse:
      type: object
      properties:
        chain:
          type: integer
        store:
          type: integer
        store_name:
          type: string
        pos:
          type: integer
        channel_POS:
          type: string
        payment_provider_code:
          type: integer
        error:
          $ref: '#/components/schemas/ErrorInfo'
    MerchantEventsResponse:
      type: object
      description: Respuesta normalizada de merchant events.